sm8450-common: Set BOARD_EXCLUDE_KERNEL_FROM_RECOVERY_IMAGE
According to [1], if the device is A/B and has a recovery partition,
which in our case, is Virtual A/B and has a dedicated recovery partition
for each slot.

Exclude the kernel image from the recovery partition, since our bootloader
doesn't support it. This will stop the bootloader from showing "Your device
is corrupted" when entering recovery, while the real problem is just because
of our recovery image has a kernel image inside it.

[1]: https://source.android.com/docs/core/architecture/partitions/generic-boot#option-2

Change-Id: I40b4c352e5308d1d3fc8efa0c86af9f10d1ded2c